# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AppCloud API",
    "description": "Knowledge graph platform for infrastructure dependency mapping and blast-radius analysis.",
    "version": "1.1.0"
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
...
  },
  "paths": {
    "/applications/": {
      "get": {
        "summary": "List Applications",
        "tags": [
          "Applications"
        ],
        "description": "Returns every Application in the graph with a `componentCount` of contained Components, ordered by tier then name.",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true,
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Stable Application UUID"
                      },
                      "name": {
                        "type": "string",
                        "example": "Payments Platform"
                      },
                      "tier": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 3,
                        "description": "1 = critical, 3 = best-effort"
                      },
                      "owner": {
                        "type": "string",
                        "example": "payments-team"
                      },
                      "environment": {
                        "type": "string",
                        "example": "production"
                      },
                      "availability": {
                        "type": "string",
                        "example": "99.99"
                      },
                      "confidentiality": {
                        "type": "string",
                        "example": "confidential"
                      },
                      "domain": {
                        "type": "string",
                        "example": "finance"
                      },
                      "componentCount": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create an Application",
        "tags": [
          "Applications"
        ],
        "description": "Creates a new Application node. `tier` is required; reasonable defaults are filled in for `availability` (`99.9`) and `confidentiality` (`internal`).",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "tier"
                ],
                "additionalProperties": true,
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "tier": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 3
                  },
                  "owner": {
                    "type": "string"
                  },
                  "environment": {
                    "type": "string"
                  },
                  "availability": {
                    "type": "string"
                  },
                  "confidentiality": {
                    "type": "string"
                  },
                  "domain": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "description": "Stable Application UUID"
                    },
                    "name": {
                      "type": "string",
                      "example": "Payments Platform"
                    },
                    "tier": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 3,
                      "description": "1 = critical, 3 = best-effort"
                    },
                    "owner": {
                      "type": "string",
                      "example": "payments-team"
                    },
                    "environment": {
                      "type": "string",
                      "example": "production"
                    },
                    "availability": {
                      "type": "string",
                      "example": "99.99"
                    },
                    "confidentiality": {
                      "type": "string",
                      "example": "confidential"
                    },
                    "domain": {
                      "type": "string",
                      "example": "finance"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "additionalProperties": true,
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "infraId required"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "additionalProperties": true,
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "infraId required"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "additionalProperties": true,
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "infraId required"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "additionalProperties": true,
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "infraId required"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "additionalProperties": true,
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "infraId required"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "additionalProperties": true,
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "infraId required"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/applications/{id}/topology": {
      "get": {
        "summary": "Application topology — Components, Infra, intra-app connections",
        "tags": [
          "Applications"
        ],
        "description": "Returns the Application, every Component it contains, the Infra each Component is deployed on, and Component↔Component `:CONNECTS_TO` edges (with protocol/port).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Resource UUID or unique name"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "properties": {
                    "application": {
                      "type": "object",
                      "additionalProperties": true,
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "description": "Stable Application UUID"
                        },
                        "name": {
                          "type": "string",
                          "example": "Payments Platform"
                        },
                        "tier": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 3,
                          "description": "1 = critical, 3 = best-effort"
                        },
                        "owner": {
                          "type": "string",
                          "example": "payments-team"
                        },
                        "environment": {
                          "type": "string",
                          "example": "production"
                        },
                        "availability": {
                          "type": "string",
                          "example": "99.99"
                        },
                        "confidentiality": {
                          "type": "string",
                          "example": "confidential"
                        },
                        "domain": {
                          "type": "string",
                          "example": "finance"
                        }
                      }
                    },
                    "components": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true,
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "name": {
                            "type": "string",
                            "example": "payment-api"
                          },
                          "type": {
                            "type": "string",
                            "example": "api",
                            "description": "service / api / worker / database / queue / cache / frontend / platform"
                          },
                          "runtime": {
                            "type": [
                              "null",
                              "string"
                            ],
                            "example": "nodejs"
                          },
                          "appId": {
                            "type": [
                              "null",
                              "string"
                            ],
                            "description": "Containing Application id (or null if unattached)"
                          },
                          "appName": {
                            "type": [
                              "null",
                              "string"
                            ]
                          }
                        }
                      }
                    },
                    "infra": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true,
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "cloud_id": {
                            "type": [
                              "null",
                              "string"
                            ],
                            "description": "Cloud-native identifier — ARN / ARM ID / GCP self-link"
                          },
                          "name": {
                            "type": "string"
                          },
                          "provider": {
                            "type": "string",
                            "enum": [
                              "aws",
                              "azure",
                              "gcp"
                            ]
                          },
                          "resource_type": {
                            "type": "string",
                            "example": "vm"
                          },
                          "region": {
                            "type": "string",
                            "example": "eastus"
                          },
                          "public": {
                            "type": "boolean"
                          },
                          "tags": {
                            "type": [
                              "object",
                              "string"
                            ],
                            "description": "Free-form tag map (object) or stringified JSON (legacy)"
                          },
                          "raw": {
                            "type": [
                              "object",
                              "string"
                            ],
                            "description": "Provider-specific payload"
                          }
                        }
                      }
                    },
                    "connections": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "additionalProperties": true,
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "infraId required"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/applications/{id}/dependencies": {
      "get": {
        "summary": "Cross-application dependencies for one Application",
        "tags": [
          "Applications"
        ],
        "description": "Walks each Component in the Application and lists Components in *other* Applications it has a `:CONNECTS_TO` edge to. Used by the impact-radius views.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Resource UUID or unique name"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "dependencies": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "http://appcloud.local"
    }
  ],
  "security": [
    {
      "ApiKey": []
    }
  ]
}